Create README.md
This contract have public variables that store the details about my coin (Token Name, Token Abbrv., Total Supply) 2. The contract have a mapping of addresses to balances (address => uint) 3. The contract have a mint function that takes two parameters: an address and a value. The function then increases the total supply by that number and increases the balance of the “sender” address by that amount 4. The contract have a burn function, which works the opposite of the mint function, as it destroy tokens. It take an address and value just like the mint functions. It then deduct the value from the total supply and from the balance of the “sender”. 5. Lastly, the burn function have conditionals to make sure the balance of "sender" is greater than or equal to the amount that is supposed to be burned.